requirements.txt
# Core dependencies for lightweight GraphRAG pipeline
networkx>=3.0
numpy>=1.24.0
tiktoken>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
pyvis>=0.3.2

# Community detection
python-louvain>=0.16

# Text processing
nltk>=3.8.0

# Optional: for better entity extraction
spacy>=3.7.0

# Development
pytest>=7.4.0
jupyter>=1.0.0
